A3

+++ Assignments/A3/Q1/BST.cpp
@@ -97,7 +97,9 @@ using std::nothrow;
    BST::~BST() {
 
       // to do
-      deleteBST(root);
+      if(root) {
+        deleteBST(root);
+      }
    }
